Clean up fitnesse.html.HtmlUtil.
Move reporting specific (javascript) snippets to reporting package.
